adjust some code
